Changed the constructor for the Order object to include a boolean variable that holds true if the customer has payed for their order.
I have also printed the orders on the new tab "Tables left to pay" which
displays the orders once they have arrived to the customers tables.

I have also changed the constructor and methods within
PendingOrderViewItem to contain another "payed" variable which prints
the status of the order on screen.